impl FireShockWave
